Move Config/Language to Renderer/Language
